fix type int

+++ common/common.go
@@ -184,8 +184,6 @@ func ChunkInt32(slices []int32, chunkSize int) (chunks [][]int32) {
 	return append(chunks, slices)
 }
 
-
-
 func StrContains(s []string, e string) bool {
 	for _, a := range s {
 		if a == e {
@@ -202,7 +200,7 @@ func I32Contains(s []int32, e int32) bool {
 	}
 	return false
 }
-func I64Contains(s []int32, e int32) bool {
+func I64Contains(s []int64, e int64) bool {
 	for _, a := range s {
 		if a == e {
 			return true
